requirements.txt
aio-pika==9.4.2
aiormq==6.8.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.10.0
asgi-lifespan==2.1.0
asgiref==3.9.1
babel==2.17.0
bcrypt==3.2.2
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
googleapis-common-protos==1.59.1
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.0.0
iniconfig==2.1.0
multidict==6.6.3
opentelemetry-api==1.36.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pluggy==1.6.0
propcache==0.3.2
protobuf==4.25.8
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pytest==8.3.2
pytest-asyncio==0.23.7
python-dotenv==1.1.1
python-jose==3.3.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
thrift==0.22.0
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
